Added a link to Aivika for .NET
Hi! 

Please add a link to my simulation library Aivika for .NET. This is an F# port of my Haskell simulation library of the same name. It supports discrete event simulation, system dynamics and agent-based modeling. The library is released under a dual-license model: GPLv3 and Commercial. There is the PDF documentation, but it mostly describes a commercial version, although the engine itself is fully licensed under dual GPL/Commercial. The commercial version contains only additional modules that are needed for analysis of the results.

Thanks,
David Sorokin