Update index.md
I want to share FSharp.Plotly with the awesome F# community. It is a charting library on top of plotly.js that I am supporting for two years now. It comes with examples and a Nuget package.
As this is my first step here, I hope I did everything correctly.
Thanks a lot for your great work with F# ...